Added omp_task_imp_firstprivate and omp_task_imp_shared.  Modified omp_task_private and omp_task_if.
omp_task_private: 

Modifications: <ompts:check>sum = 0;</ompts:check> and 
shared(result,known_sum) were added. Although, these variables are shared by
default, we explictly make them shared in case an implementation is
nonconformant with regards to default data environments.  sum = 0 was
added to initiallize the private copy of sum.


omp_task_if: 

Originally inside the explicit task read .result = (1 ==
count);. which would retrun a false value for the normal test. Also,
we made count and result shared for all test cases to ensure that they
are actually shared.


NEW TESTS

omp_task_imp_shared:

Purpose: to test the implicit shared of a task inside a parallel region

Description: This test uses the variable i, k, and result. There is a
single region inside a parallel region. Inside the single region is a
for loop creating task each turn of the loop. Inside the task region
the variable i is incremented as apart of an atomic statement. After
the parallel region result is set equal the value of i. Then, if
result equals NUM_TASK(predefined) then the function return true.

Normal Test: For the normal test, the variable i will be shared and so
result will be set to the variable i and the function will return
true.

Cross Test: For the crosscheck the variable i is firstprivate in the
explicit task region. Therefore, each task is using a private copy of
the variable i initialized to equal the global copy of the variable
i. The variable result will equal 0 and the function will return
false.

omp_task_imp_firstprivate:

Purpose: checks the implicit firstprivate of an explicit task inside a
parallel region

Description: The variables i, k, result, and task_result are used for
this test. The variable i is declared to be firstprivate at the
parallel region and the variables result and task result are declared
to be shared at the explicit task region. There is a parallel region
and inside the parallel region is a single region. Inside the single
region, here is a for loop that makes task each turn of the loop. In
addition, there is a for loop inside each task incrementing the
variable i. At the beginning of each task, if the variable i does not
equal 5 then task_result is set equal to 0.

Normal Test: For the normal test, by making the variable i
firstprivate at the parallel region, each task created will have a
firstprivate copy of the variable i; thus, when the task is inside the
for loop incrementing the variable i, i will only change for the
individual task and not for other tasks. The variable result is equal
to the truth statement (task_result and i==5). Thus, task_result will
remain a true value and the copy of i inside the parallel region will
equal 5. So, the function will return true.

Cross Test: For the crosscheck, every task created shares the variable
i; therefore, when all task complete, i will not equal to 5 and result
will be set equal to a false value. So, the function will return
false. The variables result and shared_result are made explicitly
private so to ensure that they indeed are shared.