Add tests for the omp masked directive
The omp_masked directive is new in OpenMP 5.1 and deprecates the
omp_master directive.  The tests check the basic functionality, i.e.
that an omp_maksed block is only executed by the primary thread with
number 0 in the team by default and that the filter clause can be used
to select a differing thread.

Signed-off-by: Christoph Niethammer <niethammer@hlrs.de>